+++ flutter_local_notifications/android/src/main/java/com/dexterous/flutterlocalnotifications/FlutterLocalNotificationsPlugin.java
@@ -868,7 +868,16 @@ public class FlutterLocalNotificationsPlugin
         checkCanScheduleExactAlarms(alarmManager);
         AlarmManagerCompat.setAlarmClock(alarmManager, epochMilli, pendingIntent, pendingIntent);
       } else {
-        alarmManager.set(AlarmManager.RTC_WAKEUP, epochMilli, pendingIntent);
+        if (notificationDetails.inexactWindowLengthMillis > 0L
+            && VERSION.SDK_INT >= VERSION_CODES.KITKAT) {
+          alarmManager.setWindow(
+              AlarmManager.RTC_WAKEUP,
+              epochMilli,
+              notificationDetails.inexactWindowLengthMillis,
+              pendingIntent);
+        } else {
+          alarmManager.set(AlarmManager.RTC_WAKEUP, epochMilli, pendingIntent);
+        }
       }
     }
   }