+++ flutter_local_notifications/android/src/main/java/com/dexterous/flutterlocalnotifications/models/NotificationDetails.java
@@ -130,6 +130,7 @@ public class NotificationDetails implements Serializable {
   private static final String CUSTOM_LAYOUT_LEGACY_NAME = "customLayoutLegacyName";
   private static final String CUSTOM_LAYOUT_COLLAPSED_NAME = "customLayoutCollapsedName";
   private static final String CUSTOM_LAYOUT_EXPANDED_NAME = "customLayoutExpandedName";
+  private static final String INEXACT_WINDOW_LENGTH_MILLIS = "inexactWindowLengthMillis";
 
   public Integer id;
   public String title;
@@ -205,6 +206,7 @@ public class NotificationDetails implements Serializable {
   // Note: this is set on the Android to save details about the icon that should be used when
   // re-hydrating scheduled notifications when a device has been restarted.
   public Integer iconResourceId;
+  public Long inexactWindowLengthMillis;
 
   public static NotificationDetails from(Map<String, Object> arguments) {
     NotificationDetails notificationDetails = new NotificationDetails();
@@ -240,6 +242,13 @@ public class NotificationDetails implements Serializable {
     }
     if (arguments.containsKey(DAY)) {
       notificationDetails.day = (Integer) arguments.get(DAY);
+    }
+
+    if (arguments.containsKey(INEXACT_WINDOW_LENGTH_MILLIS)) {
+      notificationDetails.inexactWindowLengthMillis =
+          (Long) arguments.get(INEXACT_WINDOW_LENGTH_MILLIS);
+    } else {
+      notificationDetails.inexactWindowLengthMillis = 0L;
     }
 
     readPlatformSpecifics(arguments, notificationDetails);