+++ flutter_local_notifications/lib/src/platform_specifics/android/notification_details.dart
@@ -151,7 +151,7 @@ class AndroidNotificationDetails {
       'AndroidNotificationDetails.customLayoutCollapsedName, '
       'AndroidNotificationDetails.customLayoutExpandedName for 31+',
     )
-        String? customLayoutName,
+    String? customLayoutName,
     this.subText,
     this.tag,
     this.actions,
@@ -161,6 +161,7 @@ class AndroidNotificationDetails {
     String? customLayoutLegacyName,
     this.customLayoutCollapsedName,
     this.customLayoutExpandedName,
+    this.inexactWindowLengthMillis,
   }) : customLayoutLegacyName = customLayoutLegacyName ?? customLayoutName;
 
   /// The icon that should be used when displaying the notification.
@@ -452,4 +453,9 @@ class AndroidNotificationDetails {
   /// The widget id containing the image must be named `push_image`.
   /// Use with versions SDK since S (31).
   final String? customLayoutExpandedName;
+
+  /// Length of notification window for inexact alarm in millis, if not null or
+  /// zero, inexact alarm will plan in this length, should be 10 minutes
+  /// or more, see more in docs https://developer.android.com/about/versions/14/changes/schedule-exact-alarms#use-cases
+  final int? inexactWindowLengthMillis;
 }